requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
black==23.12.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
dnspython==2.4.2
frozenlist==1.4.1
idna==3.6
isort==5.13.2
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pymongo==4.6.1
PyPDF2==3.0.1
pyright==1.1.344
requests==2.31.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
